general was_solver interface
