+++ ott/core/__init__.py
@@ -28,7 +28,7 @@ from . import implicit_differentiation
 from . import momentum
 from . import sinkhorn
 from . import sinkhorn_lr
-from . import neuraldual
+#from . import neuraldual
 from .implicit_differentiation import ImplicitDiff
 from .problems import LinearProblem
 from .sinkhorn import Sinkhorn
