+++ tests/core/continuous_barycenter_test.py
@@ -19,14 +19,12 @@ from absl.testing import absltest
 from absl.testing import parameterized
 import jax
 import jax.numpy as jnp
-import jax.test_util
 from ott.geometry import pointcloud
 from ott.core import bar_problems
 from ott.core import continuous_barycenter
 
 
-@jax.test_util.with_config(jax_numpy_rank_promotion='allow')
-class Barycenter(jax.test_util.JaxTestCase):
+class Barycenter(parameterized.TestCase):
 
   def setUp(self):
     super().setUp()
@@ -36,27 +34,36 @@ class Barycenter(jax.test_util.JaxTestCase):
     self.rng, *rngs = jax.random.split(self.rng, 3)
     b = jax.random.uniform(rngs[1], (self._num_points,))
     self._b = b / jnp.sum(b)
-
-  def test_euclidean_barycenter(self):
+  
+  @parameterized.product(
+      rank=[-1, 13],
+      epsilon=[1e-2])
+  def test_euclidean_barycenter(self, rank, epsilon):
+    print('Rank: ', rank, 'Epsilon: ', epsilon)
     rngs = jax.random.split(self.rng, 2)    
-    y1 = 2 * jax.random.uniform(rngs[0], (self._num_points, self._dim)) - 1
-    y2 = jax.random.uniform(rngs[0], (self._num_points, self._dim)) + 2
+    y1 = jax.random.uniform(rngs[0], (self._num_points, self._dim))
+    y2 = jax.random.uniform(rngs[1], (self._num_points, self._dim)) + 2
     y = jnp.concatenate((y1, y2))
     bar_prob = bar_problems.BarycenterProblem(
       y, num_per_segment=jnp.array([32, 29, 23, 29, 25, 29, 27, 22]),
       num_segments=8,
-      max_measure_size=50)
+      max_measure_size=40)
     threshold = 1e-3
+    kwargs = {'gamma' : 1, 'jit': False} if rank > 0 else {}
     solver = continuous_barycenter.WassersteinBarycenter(
-      epsilon=.001,
-      threshold = threshold)
-    out = solver(bar_prob, bar_size=19)
+      epsilon=epsilon,
+      rank=rank,
+      threshold = threshold,
+      **kwargs)
+    out = solver(bar_prob, bar_size=31)
     costs = out.costs
+    print('COSTS', costs)
     costs = costs[costs > -1]
-    print(costs)
-    print(out.x)
-    print(out.linear_convergence)
+    print('COSTS', costs)
     self.assertTrue(jnp.isclose(costs[-2], costs[-1], rtol=threshold))
+    self.assertTrue(jnp.isclose(costs[-2], costs[-1], rtol=threshold))
+    self.assertTrue(jnp.all(out.x.ravel()<2))
+    self.assertTrue(jnp.all(out.x.ravel()>1))
 
 if __name__ == '__main__':
   absltest.main()
